Conflicts:
	safe_qgis/plugin.py

+++ safe/common/utilities.py
@@ -53,6 +53,9 @@ def temp_dir(sub_dir='work'):
        print tmpfile
        /tmp/inasafe/23-08-2012/timlinux/testing/tmpMRpF_C
 
+    If you specify INASAFE_WORK_DIR as an environment var, it will be
+    used in preference to the system temp directory.
+
     Args:
         sub_dir str - optional argument which will cause an additional
                 subirectory to be created e.g. /tmp/inasafe/foo/
@@ -65,12 +68,17 @@ def temp_dir(sub_dir='work'):
     """
     user = getpass.getuser().replace(' ', '_')
     current_date = date.today()
-    date_string = current_date.strftime("%d-%m-%Y")
-    # Following 4 lines are a workaround for tempfile.tempdir() unreliabilty
-    handle, filename = mkstemp()
-    os.close(handle)
-    new_directory = os.path.dirname(filename)
-    os.remove(filename)
+    date_string = current_date.isoformat()
+    if 'INASAFE_WORK_DIR' in os.environ:
+        new_directory = os.environ['INASAFE_WORK_DIR']
+    else:
+        # Following 4 lines are a workaround for tempfile.tempdir()
+        # unreliabilty
+        handle, filename = mkstemp()
+        os.close(handle)
+        new_directory = os.path.dirname(filename)
+        os.remove(filename)
+
     path = os.path.join(new_directory, 'inasafe', date_string, user, sub_dir)
 
     if not os.path.exists(path):
@@ -112,6 +120,9 @@ def unique_filename(**kwargs):
 
     if 'dir' not in kwargs:
         path = temp_dir('impacts')
+        kwargs['dir'] = path
+    else:
+        path = temp_dir(kwargs['dir'])
         kwargs['dir'] = path
     if not os.path.exists(kwargs['dir']):
         # Ensure that the dir mask won't conflict with the mode
